refactor: extract library-open lifecycle into sibling file (step 9)
Move show_setup_window, on_setup_complete, and open_library off the
impl MomentsApplication block in mod.rs into a new lifecycle.rs, along
with the setup_failure_dialog, open_failure_dialog, and
show_library_error_dialog helpers (all used only by those methods).

Pure code movement — no behavior changes. The three lifecycle methods
gain pub(in crate::application) visibility so the imp::activate and
startup::start call sites still resolve them across module boundaries;
the dialog helpers stay private to lifecycle.rs. mod.rs drops from 530
to 369 lines.
